Move pricecode into separate folder
