+++ src/constants/PriceCode.ts
@@ -1,0 +1,7 @@
+export const PriceCode = {
+  REGULAR: "REGULAR",
+  NEW_RELEASE: "NEW_RELEASE",
+  CHILDREN: "CHILDREN",
+} as const;
+
+export type PriceCode = (typeof PriceCode)[keyof typeof PriceCode];
